# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Flora API",
    "version": "1.0.0",
    "description": "API для колекції букетів Flora"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://myApp.onrender.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/api/bouquets": {
      "get": {
        "tags": ["Bouquets"],
        "summary": "Отримати список усіх букетів",
        "responses": {
          "200": {
            "description": "Успішно",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Bouquet" }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Bouquets"],
        "summary": "Створити новий букет",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BouquetCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Створено",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Bouquet" }
              }
            }
          },
          "400": { "description": "Помилка валідації" }
        }
      }
    },
...
    "/api/orders": {
      "post": {
        "tags": ["Orders"],
        "summary": "Створити нове замовлення",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OrderCreate" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Створено",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "400": { "description": "Помилка валідації" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Bouquet": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid", "example": "51522bb0-0ff1-43d4-821b-e96741b4d1ab" },
          "title": { "type": "string", "example": "Rose Bouquet" },
          "description": { "type": "string", "example": "test description" },
          "price": { "type": "number", "example": 500 },
          "photoURL": { "type": "string", "example": "https://example.com/photo.jpg" },
          "favorite": { "type": "boolean", "example": false },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "BouquetCreate": {
        "type": "object",
        "required": ["title", "description", "price"],
        "properties": {
          "title": { "type": "string", "example": "Rose Bouquet" },
          "description": { "type": "string", "example": "Beautiful red roses" },
          "price": { "type": "number", "example": 500 },
          "photoURL": { "type": "string", "example": "https://example.com/photo.jpg" },
          "favorite": { "type": "boolean", "example": false }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string", "example": "Олена" },
          "phone": { "type": "string", "example": "+380501234567" },
          "address": { "type": "string", "example": "Київ, вул. Шевченка 1" },
          "message": { "type": "string", "example": "Доставити до 18:00" },
          "bouquetId": { "type": "string", "format": "uuid" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "OrderCreate": {
        "type": "object",
        "required": ["name", "phone", "address", "bouquetId"],
        "properties": {
          "name": { "type": "string", "example": "Олена" },
          "phone": { "type": "string", "example": "+380501234567" },
          "address": { "type": "string", "example": "Київ, вул. Шевченка 1" },
          "message": { "type": "string", "example": "Доставити до 18:00" },
          "bouquetId": { "type": "string", "format": "uuid" }
        }
      },
...
    }
  }
}